refactor(webui): map-viewer.js のアイコン/扇形描画を分離 (#346)
マーカー/route アイコン生成を map-icons.js へ、POI 扇形描画と yaw
回転ハンドルを MapViewerSector (map-viewer-sector.js) へ切り出し。
MapViewer は単一クラスのまま、挙動不変。
